+++ sentry/src/test/java/io/sentry/transport/HttpTransportTest.kt
@@ -1,19 +1,23 @@
 package io.sentry.transport
 
 import com.nhaarman.mockitokotlin2.any
+import com.nhaarman.mockitokotlin2.check
 import com.nhaarman.mockitokotlin2.eq
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.never
 import com.nhaarman.mockitokotlin2.verify
+import com.nhaarman.mockitokotlin2.verifyZeroInteractions
 import com.nhaarman.mockitokotlin2.whenever
 import io.sentry.ISerializer
 import io.sentry.SentryEnvelope
 import io.sentry.SentryEvent
 import io.sentry.SentryOptions
+import io.sentry.SentryOptions.Proxy
 import io.sentry.Session
 import io.sentry.protocol.User
 import java.io.IOException
-import java.net.Proxy
+import java.net.InetSocketAddress
+import java.net.Proxy.Type
 import java.net.URI
 import java.net.URL
 import javax.net.ssl.HostnameVerifier
@@ -22,6 +26,7 @@ import javax.net.ssl.SSLSocketFactory
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
+import kotlin.test.assertNull
 import kotlin.test.assertTrue
 
 class HttpTransportTest {
@@ -35,6 +40,7 @@ class HttpTransportTest {
         var readTimeout = 500
         val connection = mock<HttpsURLConnection>()
         val currentDateProvider = mock<ICurrentDateProvider>()
+        val authenticatorWrapper = mock<AuthenticatorWrapper>()
         var sslSocketFactory: SSLSocketFactory? = null
         var hostnameVerifier: HostnameVerifier? = null
 
@@ -52,7 +58,7 @@ class HttpTransportTest {
             options.sslSocketFactory = sslSocketFactory
             options.hostnameVerifier = hostnameVerifier
 
-            return object : HttpTransport(options, requestUpdater, connectionTimeout, readTimeout, sslSocketFactory, hostnameVerifier, dsn, currentDateProvider) {
+            return object : HttpTransport(options, requestUpdater, connectionTimeout, readTimeout, sslSocketFactory, hostnameVerifier, dsn, currentDateProvider, authenticatorWrapper) {
                 override fun open(): HttpsURLConnection {
                     return connection
                 }
@@ -311,6 +317,40 @@ class HttpTransportTest {
         transport.send(createEnvelope())
 
         verify(fixture.connection, never()).hostnameVerifier = any()
+    }
+
+    @Test
+    fun `When Proxy host and port are given, set to connection`() {
+        fixture.proxy = Proxy("proxy.example.com", "8090")
+        val transport = fixture.getSUT()
+
+        transport.send(createEnvelope())
+
+        assertEquals(java.net.Proxy(Type.HTTP, InetSocketAddress("proxy.example.com", 8090)), transport.proxy)
+    }
+
+    @Test
+    fun `When Proxy username and password are given, set to connection`() {
+        fixture.proxy = Proxy("proxy.example.com", "8090", "some-user", "some-password")
+        val transport = fixture.getSUT()
+
+        transport.send(createEnvelope())
+
+        verify(fixture.authenticatorWrapper).setDefault(check<ProxyAuthenticator> {
+            assertEquals("some-user", it.user)
+            assertEquals("some-password", it.password)
+        })
+    }
+
+    @Test
+    fun `When Proxy port has invalid format, proxy is not set to connection`() {
+        fixture.proxy = Proxy("proxy.example.com", "xxx")
+        val transport = fixture.getSUT()
+
+        transport.send(createEnvelope())
+
+        assertNull(transport.proxy)
+        verifyZeroInteractions(fixture.authenticatorWrapper)
     }
 
     private fun createSession(): Session {